fix null in listener list, fix maptext lines not removing themselves from lines list.

+++ code/modules/speech/floating_chat.dm
@@ -35,6 +35,11 @@
 		if(istype(src.loc, /obj/chat_maptext_holder))
 			var/obj/chat_maptext_holder/holder = src.loc
 			holder.lines -= src
+		if(istype(src.loc, /obj/item/organ/head))
+			var/obj/item/organ/head/head =  src.loc
+			var/obj/chat_maptext_holder/holder = head.linked_human?.chat_text
+			if(holder)
+				holder.lines -= src
 		for(var/client/C in src.visible_to)
 			C.images -= src
 		src.loc = null
@@ -80,11 +85,11 @@ proc/make_chat_maptext(atom/target, msg, style = "", alpha = 255, force = 0, tim
 
 	if (istype(target,/obj/item/organ/head)) // skeleton speech scrolling bandaid fix
 		var/obj/item/organ/head/head = target
-		holder = head.linked_human.chat_text
+		holder = head.linked_human?.chat_text
 		text.loc = head
 		text.layer = head.layer+1 // dont layer under the hud k thanks bye
 
-	else if(istype(target, /atom/movable))
+	else if(istype(target, /atom/movable) && holder)
 		var/atom/movable/L = target
 		holder = L.chat_text
 		text.loc = holder
@@ -92,12 +97,13 @@ proc/make_chat_maptext(atom/target, msg, style = "", alpha = 255, force = 0, tim
 	else
 		text.loc = target
 
-	if(length(holder.lines) && holder.lines[length(holder.lines)].maptext == text.maptext)
-		holder.lines[length(holder.lines)].transform *= 1.05
-		qdel(text)
-		return null
+	if (holder)
+		if(length(holder.lines) && holder.lines[length(holder.lines)].maptext == text.maptext)
+			holder.lines[length(holder.lines)].transform *= 1.05
+			qdel(text)
+			return null
 
-	holder.lines.Add(text)
+		holder.lines.Add(text)
 
 	animate(text, alpha = alpha, maptext_y = 34, time = 4, flags = ANIMATION_END_NOW)
 	var/text_id = text.unique_id
